backend/requirements.txt
mysql-connector-python==9.2.0
Flask==2.3.3